requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1

# Admin Interface (optional)
Flask-Admin==1.6.1

# Database Drivers
PyMySQL==1.1.0
# pyodbc==5.0.1  # Uncomment for SQL Server support (requires ODBC driver)

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
